Final project, only classification with svm
