+++ Assignment1/README.md
@@ -1,0 +1,47 @@
+# EBC 7100 Assignment 1
+Assignment 1 EBC 7100 text classification using supervised algorithms
+=======
+Authors:
+1. Jiaxi Chen (jchen421@uottawa.ca)
+2. Salauddin Ali Ahmed (salia023@uottawa.ca)
+3. Vrushti Buch (vbuch104@uottawa.ca)
+
+The attached source code only contains one python script.
+Pre-requisites:
+1. Python 3
+2. Any python IDE (we used Spider and Jupyter notebooks using Anaconda)
+
+Please note that this source code does not require any dependencies or environment variables to be set.
+
+The code structure is as follows:
+
+1. Importing libraries 
+2. User defined functions
+3. Data Preparation and Pre-processing
+4. Vectorization
+5. Cross Validation
+6. Error Analysis
+
+Steps to execute the script:
+1. Install Python 3 (https://www.python.org/download/releases/3.0/) or use homebrew if you are a mac user
+2. Ensure Python 3 installed properly using the following command 
+3. install - python
+4. To test if Python is successfully installed, type python on terminal/console. 
+   "You should see the output as below:
+    Python 3.7.2 (default, Dec 29 2018, 00:00:04) 
+    [Clang 4.0.1 (tags/RELEASE_401/final)] :: Anaconda, Inc. on darwin
+    Type "help", "copyright", "credits" or "license" for more information.
+    >>>" 
+6. Download the "assignment1".py file to your local machine
+7. Open command prompt (Windows) or terminal (macintosh)
+8. Enter the command - "python <path_of_the_downloaded py file>"
+9. For the best results and proper vizualization of charts, we recommend using Spider or Jupyter notebook.
+   Download the program and open it with Jupyter notebook and run the cell.  
+
+Deliverables:
+1. Project Report
+2. Source Code along with the ReadMe File.
+3. Results in HTML format
+4. Presentation
+
+For further questions, feel free to reach us at the e-mail address provided above.
